refactor(utils): move line code to core.lines and match code to core.match_indices
