+++ ggshield/core/utils.py
@@ -1,163 +1,17 @@
 import logging
 import os
-import re
 from datetime import datetime
 from itertools import islice
-from typing import Iterable, List, NamedTuple, Optional, TypeVar
+from typing import Iterable, List, Optional, TypeVar
 
 from dotenv import load_dotenv
-from pygitguardian.models import Match
 
-from ggshield.utils.git_shell import Filemode, get_git_root, is_git_dir
+from ggshield.utils.git_shell import get_git_root, is_git_dir
 
-from .text_utils import Line, LineCategory, display_error
+from .text_utils import display_error
 
 
 logger = logging.getLogger(__name__)
-
-
-REGEX_PATCH_HEADER = re.compile(
-    r"^(?P<line_content>@@ -(?P<pre_index>\d+),?\d* \+(?P<post_index>\d+),?\d* @@(?: .+)?)"  # noqa
-)
-
-
-def get_lines_from_content(
-    content: str, filemode: Filemode, is_patch: bool
-) -> List[Line]:
-    """
-    Return the secrets and the lines with line number.
-
-    :param content: Content to scan
-    :param filemode: Filemode of the content
-    :param is_patch: Is the content a patch
-    """
-
-    # Patch
-    if is_patch:
-        return list(get_lines_from_patch(content, filemode))
-
-    # File
-    return list(get_lines_from_file(content))
-
-
-def get_lines_from_file(content: str) -> Iterable[Line]:
-    """Return the lines with line number from a file."""
-    for line_count, line_content in enumerate(content.split("\n")):
-        yield Line(
-            content=line_content, category=LineCategory.data, pre_index=line_count + 1
-        )
-
-
-def get_lines_from_patch(content: str, filemode: Filemode) -> Iterable[Line]:
-    """Return the lines with line number from a git patch."""
-    content += "\n"
-    pre_index = 0
-    post_index = 0
-
-    for line in content.split("\n"):
-        line_type = line[:1]
-        line_content = ""
-        line_pre_index = None
-        line_post_index = None
-        category = None
-
-        if line_type == " ":
-            line_content = line[1:]
-            pre_index += 1
-            post_index += 1
-            line_pre_index = pre_index
-            line_post_index = post_index
-        elif line_type == "@":
-            m = REGEX_PATCH_HEADER.search(line)
-            if m is None:
-                continue
-            pre_index = int(m.groupdict()["pre_index"])
-            post_index = int(m.groupdict()["post_index"])
-            line_content = m.groupdict()["line_content"][:-1]
-
-            if filemode == Filemode.NEW or filemode == Filemode.DELETE:
-                pre_index = 1
-                post_index = 1
-
-            if line_content:
-                line_type = " "
-                pre_index -= 1
-                post_index -= 1
-                line_pre_index = None
-                line_post_index = None
-                category = LineCategory.empty
-        elif line_type == "+":
-            post_index += 1
-            line_post_index = post_index
-            line_content = line[1:]
-            category = LineCategory.addition
-        elif line_type == "-":
-            pre_index += 1
-            line_pre_index = pre_index
-            line_content = line[1:]
-            category = LineCategory.deletion
-        elif line_type == "\\":
-            # This type of line shouldn't contain any secret; no need to set indices
-            line_content = line[1:]
-
-        if line_type and line_content is not None:
-            yield Line(
-                content=line_content,
-                category=category,
-                pre_index=line_pre_index,
-                post_index=line_post_index,
-            )
-
-
-class MatchIndices(NamedTuple):
-    line_index_start: int
-    line_index_end: int
-    index_start: int
-    index_end: int
-
-
-def find_match_indices(match: Match, lines: List[Line], is_patch: bool) -> MatchIndices:
-    """Utility function.
-
-    Returns a MatchIndices instance where
-     - line_index_{start,end} are the indices in the lines of the line objects
-       containing the start and end of the match
-     - index_{start,end} are the indices of the match in the line_{start,end} objects
-
-    :param match: a Match where index_{start,end} are not None
-    :param lines: List of content lines with indices (post_index and pre_index)
-    :param is_patch: True if is patch from git, False if file
-
-    :return: MatchIndices
-    """
-    index = 0
-    line_index = 0
-    len_line = len(lines[line_index].content) + 1 + int(is_patch)
-    # Update line_index until we find the secret start
-    assert match.index_start is not None
-    while match.index_start >= index + len_line:
-        index += len_line
-        line_index += 1
-        len_line = len(lines[line_index].content) + 1 + int(is_patch)
-
-    line_index_start = line_index
-    index_start = match.index_start - index - int(is_patch)
-
-    # Update line_index until we find the secret end
-    assert match.index_end is not None
-    while match.index_end > index + len_line:
-        index += len_line
-        line_index += 1
-        len_line = len(lines[line_index].content) + 1 + int(is_patch)
-
-    line_index_end = line_index
-    index_end = match.index_end - index - int(is_patch) + 1
-    return MatchIndices(
-        line_index_start,
-        line_index_end,
-        index_start,
-        index_end,
-    )
 
 
 def _find_dot_env() -> Optional[str]:
