+++ ggshield/verticals/secret/output/secret_json_output_handler.py
@@ -4,8 +4,8 @@ from pygitguardian.client import VERSIONS
 from pygitguardian.models import Match, PolicyBreak
 
 from ggshield.core.filter import censor_content, leak_dictionary_by_ignore_sha
-from ggshield.core.text_utils import Line
-from ggshield.core.utils import find_match_indices, get_lines_from_content
+from ggshield.core.lines import Line, get_lines_from_content
+from ggshield.core.match_indices import find_match_indices
 from ggshield.utils.git_shell import Filemode
 
 from ..secret_scan_collection import Error, Result, SecretScanCollection
